omi_firmware/requirements.txt
openai-whisper>=20231117
bleak>=0.21.1
numpy>=1.24.0
requests>=2.31.0
